+++ tests/misc/fallback_ast_src/php80_or_newer/002_property_union_type.php
@@ -1,0 +1,16 @@
+<?php
+class Example {
+    public static int|string $scalar;
+    public stdClass|false $ip;
+}
+
+Example::$scalar = 1;
+Example::$scalar = 'value';
+Example::$scalar = null;
+Example::$scalar = false;
+Example::$scalar = [];
+$e = new Example();
+$e->ip = new stdClass();
+$e->ip = false;
+$e->ip = true;
+$e->ip = [];